+++ ShootHouse/Assets/Scene/DylanScene/DylanScripts/IDamage.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public interface IDamage
 {
-    void takeDamage(int amount);
+    void takeDamage(float amount);
    
 }
